+++ Libplanet.Tests/Action/ActionEvaluationExtensions.cs
@@ -1,0 +1,28 @@
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Linq;
+using System.Numerics;
+using Bencodex.Types;
+using Libplanet.Action;
+
+namespace Libplanet.Tests.Action
+{
+    public static class ActionEvaluationExtensions
+    {
+        public static IImmutableDictionary<Address, IValue> GetDirtyStates(
+            this IEnumerable<ActionEvaluation> evaluations
+        ) =>
+            evaluations.Aggregate(
+                ImmutableDictionary<Address, IValue>.Empty,
+                (dirty, ev) => dirty.SetItems(ev.OutputStates.GetUpdatedStates())
+            );
+
+        public static IImmutableDictionary<(Address, Currency), BigInteger> GetDirtyBalances(
+            this IEnumerable<ActionEvaluation> evaluations
+        ) =>
+            evaluations.Aggregate(
+                ImmutableDictionary<(Address, Currency), BigInteger>.Empty,
+                (dirty, ev) => dirty.SetItems(ev.OutputStates.GetUpdatedBalances())
+            );
+    }
+}