+++ Libplanet/Action/AccountStateGetter.cs
@@ -1,3 +1,4 @@
+#nullable enable
 using Bencodex.Types;
 
 namespace Libplanet.Action
@@ -14,5 +15,5 @@ namespace Libplanet.Action
     /// its state.</param>
     /// <returns>The account state if exists.  Otherwise <c>null</c>.
     /// </returns>
-    public delegate IValue AccountStateGetter(Address address);
+    public delegate IValue? AccountStateGetter(Address address);
 }