+++ CHANGES.md
@@ -10,6 +10,17 @@ To be released.
 
  -  Added `IAction.RenderError()` and `IAction.UnrenderError()` methods.
     [[#860], [#875]]
+ -  Added methods related fungible asset states to `IAccountStateDelta`:
+    [[#861], [#900]]
+     -  `UpdatedFungibleAssetsAccounts` property
+     -  `MintAsset(Address, Currency, BigInteger)` method
+     -  `TransferAsset(Address, Address, Currency, BigInteger)` method
+     -  `BurnAsset(Address, Currency, BigInteger)` method
+     -  `GetBalance(Address, Currency)` method
+ -  Added `IAccountStateDelta.StateUpdatedAddresses` property in order to
+    distinguish state updates from asset states.  [[#861], [#900]]
+ -  Added an optional parameter `AccountBalanceGetter accountBalanceGetter` to
+    `Block<T>.EvaluateActionsPerTx()` method.  [[#861], [#900]]
  -  `BlockChain<T>.StageTransaction()` became to throw
     `InvalidTxGenesisHashException` when it takes a `Transaction<T>` from
     a heterogeneous `BlockChain<T>` with a different genesis block.
@@ -25,9 +36,14 @@ To be released.
 
 ### Added APIs
 
+ -  Added `Currency` struct.  [[#861], [#900]]
+ -  Added `AccountBalanceGetter` delegate.  [[#861], [#900]]
  -  Added `TurnClient.BindProxies()` method. [[#756], [#868]]
  -  Added `ActionEvaluation.Exception` property.  [[#860], [[#875]]]
  -  Added `InvalidTxGenesisHashException` class.  [[#796], [#878]]
+ -  Added `CurrencyPermissionException` class.  [[#861], [#900]]
+ -  Added `InsufficientBalanceException` class.  [[#861], [#900]]
+ -  Added `BlockChain<T>.GetBalance()` method.  [[#861], [#900]]
 
 ### Behavioral changes
 
@@ -39,6 +55,8 @@ To be released.
     `UnexpectedlyTerminatedActionException` directly. Instead, it records
     an exception to `ActionEvaluation`s.  [[#860], [#875]]
  -  Added `Transaction<T>.GenesisHash` property.  [[#796], [#878]]
+ -  Added `IAccountStateDelta.UpdatedAddresses` property contains
+    asset updates besides state updates.  [[#861], [#900]]
  -  `Swarm<T>` became to ignore received transaction with different
     genesis hash.  [[#796], [#878]]
  -  `Swarm<T>` became to ignore invalid `BlockHeader`s immediately.  [[#898]]
@@ -58,6 +76,7 @@ To be released.
 [#858]: https://github.com/planetarium/libplanet/issues/858
 [#859]: https://github.com/planetarium/libplanet/pull/859
 [#860]: https://github.com/planetarium/libplanet/issues/860
+[#861]: https://github.com/planetarium/libplanet/issues/861
 [#868]: https://github.com/planetarium/libplanet/pull/868
 [#871]: https://github.com/planetarium/libplanet/issues/871
 [#875]: https://github.com/planetarium/libplanet/pull/875
@@ -66,6 +85,7 @@ To be released.
 [#883]: https://github.com/planetarium/libplanet/pull/883
 [#890]: https://github.com/planetarium/libplanet/pull/890
 [#898]: https://github.com/planetarium/libplanet/pull/898
+[#900]: https://github.com/planetarium/libplanet/pull/900
 [#902]: https://github.com/planetarium/libplanet/pull/902